[operating_system] added 'is-windows-server' function to the namespace.
Adding of this function required for checking at script available to use
some of function on Windows Server platform, for example
'cygpath::get-dos-path'.
Structure OperatingSystem was hidden. Added 'operating_system_init',
'operating_system_is_windows_server' and 'platform_is_windows_server'
functions.

[tests.xml] added calls of functions from 'operating-system' and
'platform' name spaces.

[ant4c.net.framework.module] renamed source files of the module.
Added hash sums of the nuget package to the SHA3-224SUMS and MD5SUMS
files.

[CMakeLists] added workaround approach to use pugixml at some
environment ( OpenBSD 6.8 for example ) where find_package may create
issue for using direct source of the packages in case package from
system repository was not installed ( pkg_add libpugixml ).

More details can be found at the issue discussion
https://github.com/zeux/pugixml/issues/390
and pull request
https://github.com/zeux/pugixml/pull/391

+++ ant4c.net.framework.module.c
@@ -445,7 +445,7 @@ uint8_t evaluate_function(const uint8_t* function,
 
 		case metahost_runtime_:
 			if (1 != values_count ||
-				!metahost_runtime(values[0], values_lengths[0], &values_count) ||
+				!metahost_runtime(values[0], (uint8_t)values_lengths[0], &values_count) ||
 				!bool_to_string(values_count, &output_data))
 			{
 				return 0;
